Sprint2/dlt_postgres_to_snowflake/requirements.txt
psycopg2-binary
snowflake-connector-python
python-dotenv
